pyproject.toml
[project]
name = "gad"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "scipy>=1.12",
    "numpy>=1.26",
    "httpx>=0.27",
    "cryptography>=42.0",
    "pyyaml>=6.0",
    "streamlit>=1.32",
    "streamlit-ace>=0.1.1",
    "plotly>=5.20",
    "pandas>=2.2",
    "reportlab>=4.1",
    "supabase>=2.4",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]
